[feat] Create the core figure
