+++ src/assets/styles/commons.css
@@ -12,8 +12,7 @@ h2,
   line-height: 16px;
 }
 
-h2,
-.product-code {
+h2 {
   font-size: 12px;
   line-height: 16px;
 }
